@@ -3,13 +3,20 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Tiny.RestClient.Tests.Models;
+using Tiny.RestClient.ForTest.Api.Models;
 
 namespace Tiny.RestClient.ForTest.Api.Controllers
 {
     [Route("api/[controller]")]
     public class CaseController : Controller
     {
+        // GET: api/<controller>
+        [HttpGet("Pascal")]
+        public PascalResponse PascalCase()
+        {
+            return new PascalResponse() { Id = 42, ResponseData = "REP" };
+        }
+
         // GET: api/<controller>
         [HttpGet("Camel")]
         public CamelResponse CamelCase()
