@@ -1,10 +1,9 @@
-﻿using System;
-
-namespace Tiny.RestClient.Tests.Models
+﻿namespace Tiny.RestClient.ForTest.Api.Models
 {
     public class Response
     {
         public int Id { get; set; }
+
         public string ResponseData { get; set; }
     }
 }
