@@ -1,5 +1,4 @@
-﻿#if !FILEINFO_NOT_SUPPORTED
-using System.IO;
+﻿using System.IO;
 
 namespace Tiny.RestClient
 {
@@ -11,4 +10,3 @@ namespace Tiny.RestClient
         }
     }
 }
-#endif
