@@ -44,7 +44,7 @@ namespace Tiny.RestClient
         /// <param name="contentType">The content type.</param>
         /// <returns>The current request.</returns>
         IParameterRequest AddStringContent(string content, string contentType = "text/plain");
-#if !FILEINFO_NOT_SUPPORTED
+
         /// <summary>
         /// Adds file as content.
         /// </summary>
@@ -52,7 +52,6 @@ namespace Tiny.RestClient
         /// <param name="contentType">The content type.</param>
         /// <returns>The current request.</returns>
         IParameterRequest AddFileContent(FileInfo file, string contentType);
-#endif
 
         /// <summary>
         /// As a multipart data from request.
