@@ -53,7 +53,7 @@ namespace Tiny.RestClient
         /// <exception cref="System.ArgumentNullException">thrown when content is null.</exception>
         IMultiPartFromDataExecutableRequest AddContent<TContent>(TContent content, string name = null, string fileName = null, IFormatter serializer = null, ICompression compression = null)
             where TContent : class;
-#if !FILEINFO_NOT_SUPPORTED
+
         /// <summary>
         /// Adds the content.
         /// </summary>
@@ -73,6 +73,5 @@ namespace Tiny.RestClient
         /// <returns>The current request.</returns>
         /// <exception cref="System.ArgumentNullException">thrown when content is null.</exception>
         IMultiPartFromDataExecutableRequest AddFileContent(FileInfo content, string contentType);
-#endif
     }
 }
