@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace Tiny.RestClient
 {
@@ -17,9 +19,9 @@ namespace Tiny.RestClient
         public Type TypeToSerialize => typeof(T);
         public T Data { get; }
 
-        public string GetSerializedString(IFormatter serializer, Encoding encoding)
+        public Task<string> GetSerializedStringAsync(IFormatter serializer, Encoding encoding, CancellationToken cancellationToken)
         {
-            return serializer.Serialize<T>(Data, encoding);
+            return serializer.SerializeAsync<T>(Data, encoding, cancellationToken);
         }
 
         public IFormatter Serializer { get; private set; }
